Added a lot

+++ arrays.h
@@ -1,36 +1,36 @@
 
 // Returns the maximum of two values
-long max(long v1, long v2) {
-   long m = (v1 > v2) ? v1 : v2;
+double max(double v1, double v2) {
+   double m = (v1 > v2) ? v1 : v2;
    return m;
 }
 
 // Returns the maximum of an array
-long array_max(long* values, int size) {
+double array_max(double* values, int size) {
    int i;
-   long m = values[0];
+   double m = values[0];
    for (i = 0; i < size; i++)
       m = max(m, values[i]);
    return m;
 }
 
 // Returns the minimum of two values
-long min(long v1, long v2) {
-   long m = (v1 < v2) ? v1 : v2;
+double min(double v1, double v2) {
+   double m = (v1 < v2) ? v1 : v2;
    return m;
 }
 
 // Returns the minimum of an array
-long array_min(long* values, int size) {
+double array_min(double* values, int size) {
    int i;
-   long m = values[0];
+   double m = values[0];
    for (i = 0; i < size; i++)
       m = min(m, values[i]);
    return m;
 }
 
 // Returns first index of value in array, -1 if not in array
-int index_of(long* values, int size, long value) {
+int index_of(double* values, int size, double value) {
    int i;
    for (i = 0; i < size; i++)
       if (values[i] == value)
@@ -40,7 +40,7 @@ int index_of(long* values, int size, long value) {
 }
 
 // Returns the number of occurances of value in array
-int occur_of(long* values, int size, long value) {
+int occur_of(double* values, int size, double value) {
    int i = index_of(values, size, value);
    if (i < 0)
       return 0;
@@ -48,7 +48,7 @@ int occur_of(long* values, int size, long value) {
 }
 
 // Returns the maximum gap between two occurances of a value in the array
-int max_gap(long* values, int size, long value) {
+int max_gap(double* values, int size, double value) {
    int g = -1;
    int n = occur_of(values, size, value);
    int i, j;
